requirements.txt
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.3
click==8.1.8
comm==0.2.3
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
Flask==3.0.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
narwhals==2.5.0
nest-asyncio==1.6.0
numpy==2.2.6
openpyxl==3.1.5
packaging==24.2
pandas==2.3.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
plotly==6.0.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==20.3
pytz==2025.2
pyzmq==27.0.1
requests==2.32.5
retrying==1.4.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
zipp==3.21.0
